Narrow SSE servlet mapping to prevent HTTP 200 on invalid paths
The `/*` servlet mapping was causing issues with LibreChat where, when
LibreChat was configured to auto-detect the server authentication
method, it would get stuck stuck trying to access
`/.well-known/oauth-protected-resource/mcp` or
`/.well-known/oauth-protected-resource/sse`. Because the server assumed
all requests not going to `/mcp/*` or `/messages` were meant for the SSE
servlet, the server would just hold the connection open forever. And
because LibreChat failed to time-out (I assume because it kept receiving
SSE pings from the server), it would not allow the MCP server to be
added.

This change fixes that issue.